backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
psutil==7.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.6
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
